+++ apps/web/vite.config.ts
@@ -2,9 +2,6 @@ import { sveltekit } from "@sveltejs/kit/vite";
 import { defineConfig } from "vite";
 
 export default defineConfig({
-  build: {
-    minify: "esbuild",
-  },
   plugins: [sveltekit()],
   server: {
     proxy: {